split native buttons per platform
