+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/AndroidButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b6d55f4dbd64a4cc5a0edc15ac86dbfc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
