+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/CPlugin.c.meta
