+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/CppPlugin.cpp
