+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/IosButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fc4b5ab1279f14fa7b0cac2217ef75c8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
