+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/KotlinPlugin.kt.meta
