+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/NativeButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2e54f81e3c3f47dab2f9e6b23c88d38
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
