+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/NativeSupportScene.cs.meta
