+++ samples/unity-of-bugs/Assets/Scripts/NativeSupport/ObjectiveCPlugin.m.meta
