+++ samples/unity-of-bugs/Assets/Scripts/SceneButtons.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 500071f954f4948cc98f9d24b2728c5e
+guid: af17d449a00bc45ebb1767426e52f25b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
