+++ samples/unity-of-bugs/ProjectSettings/EditorBuildSettings.asset
@@ -9,7 +9,7 @@ EditorBuildSettings:
     path: Assets/Scenes/1_Bugfarm.unity
     guid: ea1c53ffa9f3c482287ee84d42ea0db4
   - enabled: 1
-    path: Assets/Scenes/2_MobileNativeSupport.unity
+    path: Assets/Scenes/2_NativeSupport.unity
     guid: b2239cfd14b1c400ea7357ae542078c1
   - enabled: 1
     path: Assets/Scenes/3_Transition.unity
